🐛 FIX: template checks

+++ zemfrog/generator.py
@@ -47,17 +47,17 @@ def g_extension(name: str):
     """
 
     validate_module_name(name)
-    if os.path.isfile(name):
-        print("The %r extension already exists" % name)
-        return
-
-    print("Creating a %r extension..." % name, end="")
     ext_dir = os.path.join(current_app.root_path, "extensions")
     old_filename = get_template("extension", "name.py")
     with open(old_filename) as fp:
         new_data = fp.read()
 
-    new_filename = os.path.join(ext_dir, name.lower() + ".py")
+    new_filename = os.path.join(ext_dir, name + ".py")
+    if os.path.isfile(new_filename):
+        print("The %r extension already exists" % name)
+        return
+
+    print("Creating a %r extension..." % name, end="")
     with open(new_filename, "w") as fp:
         fp.write(new_data)
 
@@ -72,11 +72,6 @@ def g_task(name: str):
     """
 
     validate_module_name(name)
-    if os.path.isfile(name):
-        print("The %r task already exists" % name)
-        return
-
-    print("Creating %r task..." % name, end="")
     import_name = get_import_name(current_app)
     main_app = True
     if import_name:
@@ -89,7 +84,12 @@ def g_task(name: str):
         py_t = Template(old_data)
         new_data = py_t.render(main_app=main_app, task_name=name)
 
-    new_filename = os.path.join(ext_dir, name.lower() + ".py")
+    new_filename = os.path.join(ext_dir, name + ".py")
+    if os.path.isfile(new_filename):
+        print("The %r task already exists" % name)
+        return
+
+    print("Creating %r task..." % name, end="")
     with open(new_filename, "w") as fp:
         fp.write(new_data)
 
@@ -104,11 +104,6 @@ def g_model(name: str):
     """
 
     validate_module_name(name)
-    if os.path.isfile(name):
-        print("The %r model already exists" % name)
-        return
-
-    print("Creating %r model..." % name, end="")
     import_name = get_import_name(current_app)
     main_app = True
     if import_name:
@@ -121,7 +116,12 @@ def g_model(name: str):
         py_t = Template(old_data)
         new_data = py_t.render(main_app=main_app, model_name=name.capitalize())
 
-    new_filename = os.path.join(model_dir, name.lower() + ".py")
+    new_filename = os.path.join(model_dir, name + ".py")
+    if os.path.isfile(new_filename):
+        print("The %r model already exists" % name)
+        return
+
+    print("Creating %r model..." % name, end="")
     with open(new_filename, "w") as fp:
         fp.write(new_data)
 
@@ -137,19 +137,19 @@ def g_api(name: str):
     """
 
     validate_module_name(name)
-    if os.path.isfile(name):
-        print("The %r API already exists" % name)
-        return
-
-    print("Creating API %r... " % name, end="")
     api_dir = os.path.join(current_app.root_path, "api")
     old_filename = get_template("api", "name.py")
     with open(old_filename) as fp:
         old_data = fp.read()
         py_t = Template(old_data)
-        new_data = py_t.render(name=name, url_prefix=name.lower())
+        new_data = py_t.render(name=name, url_prefix=name)
 
-    new_filename = os.path.join(api_dir, name.lower() + ".py")
+    new_filename = os.path.join(api_dir, name + ".py")
+    if os.path.isfile(new_filename):
+        print("The %r REST API already exists" % name)
+        return
+
+    print("Creating REST API %r... " % name, end="")
     with open(new_filename, "w") as fp:
         fp.write(new_data)
 
@@ -165,10 +165,6 @@ def g_api_crud(name: str):
     """
 
     validate_module_name(name)
-    if os.path.isfile(name):
-        print("The %r REST API already exists" % name)
-        return
-
     src_model = search_model(name)
     import_name = get_import_name(current_app)
     main_app = True
@@ -178,19 +174,23 @@ def g_api_crud(name: str):
         src_model = src_model[idx:]
 
     api_dir = os.path.join(current_app.root_path, "api")
-    print("Creating REST API %r... " % name, end="")
     old_filename = get_template("crud", "name.py")
     with open(old_filename) as fp:
         old_data = fp.read()
         py_t = Template(old_data)
         new_data = py_t.render(
             name=name,
-            url_prefix=name.lower(),
+            url_prefix=name,
             src_model=src_model,
             main_app=main_app,
         )
 
-    new_filename = os.path.join(api_dir, name.lower() + ".py")
+    new_filename = os.path.join(api_dir, name + ".py")
+    if os.path.isfile(new_filename):
+        print("The %r REST API already exists" % name)
+        return
+
+    print("Creating REST API %r... " % name, end="")
     with open(new_filename, "w") as fp:
         fp.write(new_data)
 
@@ -206,12 +206,12 @@ def g_blueprint(name: str):
     """
 
     validate_module_name(name)
-    if os.path.isdir(name):
+    bp_dir = os.path.join(current_app.root_path, name)
+    if os.path.isdir(bp_dir):
         print("The %r blueprint already exists" % name)
         return
 
     print("Creating blueprint %r... " % name, end="")
-    bp_dir = os.path.join(current_app.root_path, name.lower())
     copy_template("blueprint", bp_dir)
     for fname in ("routes", "urls"):
         filename = os.path.join(bp_dir, fname + ".py")
@@ -235,17 +235,17 @@ def g_middleware(name: str):
     """
 
     validate_module_name(name)
-    if os.path.isfile(name):
-        print("The %r middleware already exists" % name)
-        return
-
-    print("Creating middleware %r... " % name, end="")
     middleware_dir = os.path.join(current_app.root_path, "middlewares")
     old_filename = get_template("middleware", "name.py")
     with open(old_filename) as fp:
         data = fp.read()
 
-    new_filename = os.path.join(middleware_dir, name.lower() + ".py")
+    new_filename = os.path.join(middleware_dir, name + ".py")
+    if os.path.isfile(new_filename):
+        print("The %r middleware already exists" % name)
+        return
+
+    print("Creating middleware %r... " % name, end="")
     with open(new_filename, "w") as fp:
         fp.write(data)
 
@@ -261,11 +261,6 @@ def g_command(name: str):
     """
 
     validate_module_name(name)
-    if os.path.isfile(name):
-        print("The %r command already exists" % name)
-        return
-
-    print("Creating command %r..." % name, end="")
     cmd_dir = os.path.join(current_app.root_path, "commands")
     old_filename = get_template("command", "name.py")
     with open(old_filename) as fp:
@@ -273,7 +268,12 @@ def g_command(name: str):
         py_t = Template(old_data)
         new_data = py_t.render(name=name)
 
-    new_filename = os.path.join(cmd_dir, name.lower() + ".py")
+    new_filename = os.path.join(cmd_dir, name + ".py")
+    if os.path.isfile(new_filename):
+        print("The %r command already exists" % name)
+        return
+
+    print("Creating command %r..." % name, end="")
     with open(new_filename, "w") as fp:
         fp.write(new_data)
 
@@ -289,17 +289,17 @@ def g_error_handler(name):
     """
 
     validate_module_name(name)
-    if os.path.isfile(name):
-        print("The %r error handler already exists" % name)
-        return
-
-    print("Creating error handler %r... " % name, end="")
     eh_dir = os.path.join(current_app.root_path, "handlers")
     old_filename = get_template("errorhandler", "name.py")
     with open(old_filename) as fp:
         data = fp.read()
 
-    new_filename = os.path.join(eh_dir, name.lower() + ".py")
+    new_filename = os.path.join(eh_dir, name + ".py")
+    if os.path.isfile(new_filename):
+        print("The %r error handler already exists" % name)
+        return
+
+    print("Creating error handler %r... " % name, end="")
     with open(new_filename, "w") as fp:
         fp.write(data)
 
@@ -315,17 +315,17 @@ def g_loader(name):
     """
 
     validate_module_name(name)
-    if os.path.isfile(name):
-        print("The %r loader already exists" % name)
-        return
-
-    print("Creating a loader %r... " % name, end="")
     ld_dir = os.path.join(current_app.root_path, "loaders")
     old_filename = get_template("loader", "name.py")
     with open(old_filename) as fp:
         data = fp.read()
 
-    new_filename = os.path.join(ld_dir, name.lower() + ".py")
+    new_filename = os.path.join(ld_dir, name + ".py")
+    if os.path.isfile(new_filename):
+        print("The %r loader already exists" % name)
+        return
+
+    print("Creating a loader %r... " % name, end="")
     with open(new_filename, "w") as fp:
         fp.write(data)
